refactor controllers

+++ src/symbolic/grid_based_symbolic_model/grid_based_symbolic_model.jl
@@ -29,7 +29,7 @@ get_retained_domain(symmodel::GridBasedSymbolicModel) =
 enum_source_states(symmodel::GridBasedSymbolicModel) =
     MP.enum_states(get_source_domain(symmodel), get_state_mapping(symmodel))
 
-is_allowed_state(symmodel::GridBasedSymbolicModel, q) =
+is_allowed_state(symmodel::GridBasedSymbolicModel, q::Int) =
     q !== nothing &&
     MP.contains_state(get_retained_domain(symmodel), get_state_mapping(symmodel), q)
